requirements.txt
fastapi==0.128.8
python-multipart==0.0.20
uvicorn[standard]==0.39.0
websockets==15.0.1
deepgram-sdk==6.0.1
openai==2.29.0
pyngrok==7.5.1
python-dotenv==1.2.1
requests==2.32.5
